# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UptivaLab API
  description: |
    Comprehensive uptime monitoring and status page platform API.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Features
    - HTTP/HTTPS monitoring
    - TCP port monitoring
    - Ping monitoring
    - DNS monitoring
    - Certificate expiry tracking
    - Synthetic browser testing
    - Docker container monitoring
    - gRPC service monitoring
    - Push-based heartbeat monitoring
    - Incident management
    - Maintenance windows
    - Public status pages
    - Multi-channel notifications
    
  version: 1.0.0
  contact:
    name: UptivaLab Support
    url: https://github.com/tekuonline/uptivaLab
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.uptivalab.com/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users (admin only)
      responses:
        '200':
          description: List of users with API key counts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    email:
                      type: string
                      format: email
                    role:
                      type: string
                      enum: [ADMIN, VIEWER]
                    apiKeyCount:
                      type: number
                    createdAt:
                      type: string
                      format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required

    post:
      tags: [Users]
      summary: Create new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                  format: email
                  example: newuser@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                role:
                  type: string
                  enum: [ADMIN, VIEWER]
                  default: VIEWER
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '400':
          description: Invalid input or email already exists
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
  /users/{id}:
    delete:
      tags: [Users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Cannot delete yourself or last admin
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          description: User not found
  /users/{id}/role:
    put:
      tags: [Users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [ADMIN, VIEWER]
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
        '400':
          description: Cannot change role of last admin
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - Admin access required
        '404':
          description: User not found

  # Invitations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
...